Fix auth type issue
