+++ app/(normalLayout)/ads/[file_number]/page.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link';
 
 // Auth
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/app/api/auth/[...nextauth]/auth';
 import { redirect } from 'next/navigation';
 
 async function fetchADSDetail(file_number: string) {
