+++ app/(normalLayout)/ads/page.tsx
@@ -8,7 +8,7 @@ import Head from 'next/head';
 
 // Auth
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/app/api/auth/[...nextauth]/auth';
 import { redirect } from 'next/navigation';
 
 export default async function Home() {
